Rename 'dao' package to 'db'
